Specify target in build subcommand.

+++ lib/gamefic-sdk/build.rb
@@ -3,13 +3,15 @@ require 'yaml'
 module Gamefic::Sdk
 
   module Build
-    def self.release directory, quiet = false
+    def self.release directory, target: nil, quiet: false
       config = Gamefic::Sdk::Config.load(directory)
+      raise "Invalid target #{target}" unless target.nil? or config.targets.key?(target)
       if config.auto_import?
         puts "Importing scripts..."
         Shell.start ['import', directory, '--quiet']
       end
       config.targets.each_pair { |k, v|
+        next unless target.nil? or k == target
         plat = Gamefic::Sdk::Platform.load(config, k)
         puts "Clearing #{k}..."
         FileUtils.rm_rf plat.build_dir, secure: true
